requirements.txt
Flask>=2.0
Werkzeug>=2.0
requests>=2.20
APScheduler>=3.0
psutil>=5.0
Markdown>=3.3
boto3>=1.20
openai>=1.0.0
Flask-Babel>=3.0.0
Flask-SocketIO>=5.0.0
python-socketio>=5.0.0
python-engineio>=4.0.0
